verificato l'esito dei run fino al consumo

+++ src/sfcabm/AProductDemand.java
@@ -18,7 +18,7 @@ public class AProductDemand{
 	public void adjustDemand(int id,double factor){
 		demand=demand*factor;
 		if(Context.verboseFlag){
-			System.out.println("        consumer "+id+" revised demand of product with absolute rank "+absoluteRank+" and relative rank "+relativeRank+" demanded quantity "+demand+" previously demanded "+(int)(demand/factor));
+			System.out.println("        consumer "+id+" revised demand of product with absolute rank "+absoluteRank+" and relative rank "+relativeRank+" demanded quantity "+demand+" previously demanded "+(demand/factor));
 		}
 	}
